Cput

+++ src/gifti/gtest.h
@@ -13,7 +13,7 @@
       } while(0)
 
 int ewrite_data_line (void *, int, int, int, int, int, FILE *);
-int ewrite_many_lines(void **, int, int, int, int, FILE *);
+int ewrite_many_lines(void **, int, size_t, size_t, int, FILE *);
 int write_1D_file(DataArray ** dlist, int len, char * prefix, int add_suf);
 int write_as_ascii(gifti_image * gim, char * prefix);
 int write_surf_file( DataArray * dc, DataArray * dt, char *, int);
